+++ core/dto/settle_deliver_meta.go
@@ -9,7 +9,7 @@
 package dto
 
 type SettleDeliverMeta struct {
-	Id         int32  `db:"id"`
+	Id         int64  `db:"id"`
 	PersonName string `db:"personName"`
 	Phone      string `db:"phone"`
 	Address    string `db:"address"`
